# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BrightMart API",
        "description": "API untuk manajemen produk di BrightMart",
        "contact": {
            "email": "support@brightmart.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Display a listing of products (API).",
                "operationId": "b761016fafae7931485bfae6691bbf1b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product_name": {
                                                        "type": "string",
                                                        "example": "Laptop"
                                                    },
                                                    "product_description": {
                                                        "type": "string",
                                                        "example": "High-end laptop"
                                                    },
                                                    "product_price": {
                                                        "type": "number",
                                                        "example": 1000000
                                                    },
                                                    "product_stock": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "product_image": {
                                                        "type": "string",
                                                        "example": "123456.jpg",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "8977ad75cbfcb71382485bdd47fb8bca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "product_name",
                                    "product_description",
                                    "product_price",
                                    "product_stock"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "product_description": {
                                        "type": "string",
                                        "example": "High-end laptop"
                                    },
                                    "product_price": {
                                        "type": "number",
                                        "example": 1000000
                                    },
                                    "product_stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "product_image": {
                                        "type": "file",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/store": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "operationId": "ed30b316d51fc533775d2c5dcf0737c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "product_name",
                                    "product_description",
                                    "product_price",
                                    "product_stock"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "product_description": {
                                        "type": "string",
                                        "example": "High-end laptop"
                                    },
                                    "product_price": {
                                        "type": "number",
                                        "example": 1000000
                                    },
                                    "product_stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "product_image": {
                                        "type": "file",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product stored successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}